+++ lib/src/ui/views/settings/security/magic_starter_sessions_view.dart
@@ -51,8 +51,7 @@ class _MagicStarterSessionsViewState
 
   @override
   void onInit() {
-    controller.clearErrors();
-    controller.setEmpty();
+    controller.resetQuietly();
     if (MagicStarterConfig.hasSessionsFeatures()) {
       _loadSessions();
     }